requirements.txt
alembic==1.6.2
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
ecdsa==0.14.1
fastapi==0.65.0
fastapi-socketio==0.0.6
greenlet==1.0.0
h11==0.12.0
httptools==0.1.2
idna==2.10
iniconfig==1.1.1
jose==1.0.0
Mako==1.1.4
MarkupSafe==1.1.1
netifaces==0.10.9
packaging==20.9
pluggy==0.13.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.2.0
python-jose==3.2.0
python-socketio==5.3.0
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.6
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
